% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPrep.r
\name{DataPrep}
\alias{DataPrep}
\title{Format data to be used in regression fitting}
\usage{
DataPrep(plays, yearvec, return.ot = F)
}
\arguments{
\item{plays}{a data frame containing plays from the CFB Stats play.csv file}

\item{yearvec}{the numeric vector of years to read in the data for}

\item{return.ot}{logical indicator on what to do with overtime plays}
}
\value{
A list containing the plays in regulation. If \code{return.ot} is \code{TRUE} then the list contains a second data frame for overtime plays
}
\description{
This function takes in a data frame of plays read in from the CFB Stats
play by play file after the time of each play has been added. The function
determines a winner of each game, adjusts the down, distance, and spot
for kickoffs, determines if the game is against an FCS opponent, adds other
indicator variables, and handles overtime plays.
}
\examples{
plays <- readin("play", 2010:2014)
newtime <- AddTime(plays)
cleandata <- DataPrep(newtime, 2010:2014, return.ot = T)
head(cleandata[[1]])
head(cleandata[[2]])
}

