% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HandleOvertime.r
\name{HandleOvertime}
\alias{HandleOvertime}
\title{Structure plays correctly for games that went into overtime}
\usage{
HandleOvertime(overtime)
}
\arguments{
\item{overtime}{a data frame containing overtime plays from the CFB Stats play.csv file}
}
\value{
A data frame with the overtime plays.
}
\description{
This function takes in overtime plays and gets them in the same structure as
plays in regulation. This function is called from the DataPrep function. You
really won't need this function on it's own ever.
}
\examples{
plays <- readin("play", 2010:2014)
newtime <- AddTime(plays)
cleandata <- DataPrep(newtime, 2010:2014, return.ot = T)
head(cleandata[[1]])
head(cleandata[[2]])
}

