% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readin.r
\name{readin}
\alias{readin}
\title{Read in data from multiple years of CFB Stats data}
\usage{
readin(file, yearvec, folder_loc = "~/Documents/CFB Data/")
}
\arguments{
\item{file}{The base name of the file to read in.}

\item{yearvec}{A numeric vector of years of data to read in at once.}

\item{folder_loc}{The location on your local drive of where the CFB Stats data is stored.}
}
\value{
A data frame containing the data file for all years. An additional column has been added with the year of each data row.
}
\description{
The function accepts a base filename from the CFB Stats data files and a
vector containing the years to read the data in. If there are multiple years
to read in the years are stacked and an extra variable is added called
\code{Year} that contains the year of that data row. The package assumes
the CFB Stats data are stored in a location with different folders for
each year and the name of the folders are \code{YYYY data}, e.g.
2009 data/play.csv . The function also replaces column names with a space
with a \code{.}.
}
\examples{
plays <- readin("play", 2010:2014)
table(plays$Year)
}

